Update brochat.js
Ryan, when I was testing from mobile I noticed the web socket kept closing so I wasn't able to send or receive any messages. The reason is the ip address of the serviceLocation needs to be the ip of the machine running the chat server.

You may have already come across that but just a heads up that we'll want to change it to our ip before we demo.